error handle for an invalid comment id

+++ models/comments.models.js
@@ -80,6 +80,9 @@ exports.addCommentOnReview = (review_id, username, body) => {
 };
 
 exports.removeComment = (comment_id) => {
+  if (isNaN(+comment_id)) {
+    return notNumber("comment_id");
+  }
   return db.query(
     `
     DELETE FROM comments WHERE comment_id = $1 RETURNING *
